Stored procedures for getting available book copies, removing a book, getting books by genre, author and title.
I made the first one a procedure instead of a function because functions only return single values, but that's a select.

User Defined Functions for if a specific copy of a book is available (for use in getting available copies), and to add a copy of a book.

I made the latter a function so you can get the id of the copy you just added as a reference - could be necessary if a user has multiple copies of the same book.